Move Array2 contents into Array
