+++ src/Array.hs
@@ -1,14 +1,10 @@
-{-# LANGUAGE DefaultSignatures, FlexibleContexts, FlexibleInstances, MultiParamTypeClasses,
+{-# LANGUAGE DefaultSignatures, FlexibleContexts, FlexibleInstances, GADTs, MultiParamTypeClasses,
              TypeFamilies, UndecidableInstances #-}
 {-|
 Arrays
 
-Array type for element types isomorphic to other element types with existing array support.
-
-Primarily intended as a way to get unboxed arrays.
-As an example, if you have "newtype N = N Int", define "instance Unbox N where type Unboxed N = Int"
-and use UArr i N as an unboxed array for N. This is a lot more performant than a boxed Array i N.
-Also works with mutable arrays, IOUArr and STUArr.
+Stuff to make it easier to work with arrays from the array package, especially when it comes to
+unboxed arrays. If the vector package is available, just use that.
 
 Implementation notes:
 * The problem: Unboxed array types (UArray, STUArray, IOUArray) have the element type role as
@@ -23,7 +19,24 @@ Implementation notes:
   be Arr b arr i a and you would have to specify b when using it even though a determines b.
 * Indexing is as fast as the underlying representation but construction via listArray and array are
   known to be slower. See ArrayBench.hs. TODO: Figure out why and fix it.
+* unsafeAccum and unsafeAccumArray are optional but explicity error for Arr2 because the default
+  definitions use STArray. This is horribly slow with unboxed arrays. Better error than TLE.
 * TODO: Implement freeze and unsafeFreeze.
+
+Arr
+Array type for element types isomorphic to other element types with existing array support.
+Primarily intended as a way to get unboxed arrays.
+As an example, if you have "newtype N = N Int", define "instance Unbox N where type Unboxed N = Int"
+and use UArr i N as an unboxed array for N. This is a lot more performant than a boxed Array i N.
+Also works with mutable arrays, IOUArr and STUArr.
+
+Arr2
+Arrays for 2-tuples
+Primarily intended as a way to get unboxed arrays.
+If a and b can be put in UArrays, UArr2 i (a,b) works as an unboxed array for (a,b).
+This is a lot more performant than a boxed Array i (a,b). Also works with mutable arrays, STUArr2
+and IOUArr2.
+Can be nested to get UArr3, UArr4, etc. Use with Unbox and Arr to store your own types.
 -}
 
 module Array
@@ -32,6 +45,16 @@ module Array
     , UArr
     , IOUArr
     , STUArr
+    , Arr2
+    , UArr2
+    , UArr3
+    , UArr4
+    , STUArr2
+    , STUArr3
+    , STUArr4
+    , IOUArr2
+    , IOUArr3
+    , IOUArr4
     ) where
 
 import Control.DeepSeq
@@ -39,6 +62,9 @@ import Data.Array.Base
 import Data.Array.IO
 import Data.Coerce
 import Data.Semigroup
+
+--------------------------------------------------------------------------------
+-- Unbox and Arr
 
 class Unbox a where
     type Unboxed a
@@ -85,7 +111,48 @@ instance Unbox (Max a) where
     type Unboxed (Max a) = a
 
 --------------------------------------------------------------------------------
+-- Arr2
+
+data Arr2 arra arrb i ab where
+    Arr2 :: !(arra i a) -> !(arrb i b) -> Arr2 arra arrb i (a, b)
+
+instance (IArray arra a, IArray arrb b) => IArray (Arr2 arra arrb) (a, b) where
+    bounds (Arr2 xa _)              = bounds xa
+    numElements (Arr2 xa _)         = numElements xa
+    unsafeArray b ixys              = Arr2 (unsafeArray b (map (fmap fst) ixys)) (unsafeArray b (map (fmap snd) ixys))
+    unsafeAt (Arr2 xa ya) i         = (unsafeAt xa i, unsafeAt ya i)
+    unsafeReplace (Arr2 xa ya) ixys = Arr2 (unsafeReplace xa (map (fmap fst) ixys)) (unsafeReplace ya (map (fmap snd) ixys))
+    unsafeAccum                     = error "Arr2 unsafeAccum: not implemented"
+    unsafeAccumArray                = error "Arr2 unsafeAccumArray: not implemented"
+
+instance (IArray (Arr2 arra arrb) ab, Ix i, Show i, Show ab) => Show (Arr2 arra arrb i ab) where
+    showsPrec = showsIArray
+
+instance (Monad m, MArray marra a m, MArray marrb b m) => MArray (Arr2 marra marrb) (a, b) m where
+    getBounds (Arr2 xa _)             = getBounds xa
+    getNumElements (Arr2 xa _)        = getNumElements xa
+    unsafeNewArray_ b                 = Arr2 <$> unsafeNewArray_ b <*> unsafeNewArray_ b
+    newArray_ b                       = Arr2 <$> newArray_ b <*> newArray_ b
+    unsafeRead (Arr2 xa ya) i         = (,) <$> unsafeRead xa i <*> unsafeRead ya i
+    unsafeWrite (Arr2 xa ya) i (x, y) = unsafeWrite xa i x *> unsafeWrite ya i y
+
+type UArr2 = Arr2 UArray UArray
+type UArr3 = Arr2 UArray UArr2
+type UArr4 = Arr2 UArr2 UArr2
+
+type STUArr2 s = Arr2 (STUArray s) (STUArray s)
+type STUArr3 s = Arr2 (STUArray s) (STUArr2 s)
+type STUArr4 s = Arr2 (STUArr2 s) (STUArr2 s)
+
+type IOUArr2 = Arr2 IOUArray IOUArray
+type IOUArr3 = Arr2 IOUArray IOUArr2
+type IOUArr4 = Arr2 IOUArr2 IOUArr2
+
+--------------------------------------------------------------------------------
 -- For tests
 
 instance NFData (arr i (Unboxed a)) => NFData (Arr arr i a) where
     rnf = rnf . unArr
+
+instance (NFData (arra i a), NFData (arrb i b)) => NFData (Arr2 arra arrb i (a, b)) where
+    rnf (Arr2 xa ya) = rnf xa `seq` rnf ya
